[✨기능] detail 리팩토링
